requirements.txt
moviepy
requests
python-dotenv
google-generativeai
gTTS
langchain
langchain-google-genai
langgraph
edge-tts
streamlit
cloudinary
